アイテムが実際に処理するコード

+++ industrialization/Config/Recipe/Data/MachineRecipeData.cs
@@ -6,7 +6,7 @@ namespace industrialization.Config.Recipe.Data
     {
         public int InstallationId { get; }
 
-        public MachineRecipeData(int installationId,long time,ItemStack[] itemInputs, ItemOutput[] itemOutputs)
+        public MachineRecipeData(int installationId,int time,ItemStack[] itemInputs, ItemOutput[] itemOutputs)
         {
             InstallationId = installationId;
             ItemInputs = itemInputs;
@@ -18,7 +18,7 @@ namespace industrialization.Config.Recipe.Data
 
         public ItemOutput[] ItemOutputs { get; }
 
-        public long Time { get; }
+        public int Time { get; }
 
         public bool RecipeConfirmation(IItemStack[] inputSlot)
         {
