+++ industrialization/Config/Recipe/Data/NullMachineRecipeData.cs
@@ -7,7 +7,7 @@ namespace industrialization.Config.Recipe.Data
         public IItemStack[] ItemInputs => System.Array.Empty<IItemStack>();
         public ItemOutput[] ItemOutputs => System.Array.Empty<ItemOutput>();
         public int InstallationId => -1;
-        public long Time => 0;
+        public int Time => 0;
 
         public bool RecipeConfirmation(IItemStack[] inputSlot)
         {
